Fix tangled bindings in the composition area
We reversed the logic: instead of adding a local-map to the composition area, we
add it to the read-only thread.  This is more robust since the text of the
thread can't change.

The keymaps can be simplified: there is no more need for a separate keymap for
the linear view and the tree view.

This fixes related issues with Evil as well.